pi/requirements.txt
# DogPhone – Raspberry Pi
python-telegram-bot>=21.0
requests>=2.28.0
flask>=3.0.0
RPi.GPIO>=0.7.0; sys_platform == 'linux'
